+++ sync-for-commerce/src/sdk/models/operations/createcompany.ts
@@ -20,6 +20,12 @@ export class CreateCompanyResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * The request made is not valid.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()