+++ sync-for-commerce/src/sdk/models/operations/requestsync.ts
@@ -25,6 +25,12 @@ export class RequestSyncResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * The request made is not valid.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()